// Neptune.Web/src/app/shared/generated/model/treatment-bmp-assessment-upsert-dto.ts
/**
 * Neptune.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TreatmentBMPObservationUpsertDto } from './treatment-bmp-observation-upsert-dto';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TreatmentBMPAssessmentUpsertDto { 
    Observations?: Array<TreatmentBMPObservationUpsertDto> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TreatmentBMPAssessmentUpsertDtoForm { 
    Observations?: FormControl<Array<TreatmentBMPObservationUpsertDto>>;
}

export class TreatmentBMPAssessmentUpsertDtoFormControls { 
    public static Observations = (value: FormControlState<Array<TreatmentBMPObservationUpsertDto>> | Array<TreatmentBMPObservationUpsertDto> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<TreatmentBMPObservationUpsertDto>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}